requirements.txt
Django>=4.2.16,<=4.3
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pydantic
calendly

